+++ spec/fontcustom/watcher_spec.rb
@@ -3,7 +3,7 @@ require "fileutils"
 require "fontcustom/watcher"
 
 describe Fontcustom::Watcher do
-  # Silence messages without passing :verbose => false to everything
+  # Silence messages without passing :quiet => true to everything
   before(:each) do
     Fontcustom::Options.any_instance.stub :say_message
   end